refactor: hoist SalesByUtmWidget into widgets-toolkit (upstream pattern)
