+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/index.ts
@@ -99,6 +99,7 @@ export {
 	TotalReturnsWidget,
 	VisitorsByLocationWidget,
 	SalesByDeviceWidget,
+	SalesByUtmWidget,
 	SessionsByDeviceWidget,
 	TopPerformingProductLeaderboardWidget,
 	type TopPerformingProductLeaderboardWidgetProps,
