+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/widgets/index.ts
@@ -8,6 +8,7 @@ export {
 } from './revenue-by-customer-type';
 export { NewVsReturningCustomerWidget } from './new-vs-returning-customer';
 export { SalesByDeviceWidget } from './sales-by-device';
+export { SalesByUtmWidget } from './sales-by-utm';
 export { SessionsByDeviceWidget } from './sessions-by-device';
 export { BookingsByAttendanceWidget } from './bookings-by-attendance';
 export { TotalReturnsWidget } from './total-returns';
