+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/widgets/sales-by-utm/index.ts
@@ -1,0 +1,1 @@
+export { SalesByUtmWidget } from './sales-by-utm-widget';
