+++ projects/packages/premium-analytics/widgets/sales-by-utm-campaign/render.tsx
@@ -1,19 +1,12 @@
 /**
  * External dependencies
  */
-import { useReportOrderAttribution } from '@jetpack-premium-analytics/data';
-import { megaphone } from '@jetpack-premium-analytics/icons';
 import {
-	LeaderboardChart,
-	WidgetLoadingOverlay,
+	SalesByUtmWidget,
 	WidgetRoot,
-	buildSalesByUtmData,
-	formatLegendLabels,
-	useWidgetError,
-	useWidgetRootContext,
 	type ReportParamsFieldAttributes,
 } from '@jetpack-premium-analytics/widgets-toolkit';
-import { useMemo, type ComponentProps, type CSSProperties } from 'react';
+import { type ComponentProps } from 'react';
 /**
  * Internal dependencies
  */
@@ -29,59 +22,11 @@ type SalesByUtmCampaignRenderProps = WidgetRenderProps< SalesByUtmCampaignRender
 	setError?: ComponentProps< typeof WidgetRoot >[ 'setError' ];
 };
 
-function SalesByUtmCampaignWidget() {
-	const { reportParams } = useWidgetRootContext();
-
-	const params = useMemo(
-		() => ( {
-			...reportParams,
-			view: 'campaign' as const,
-		} ),
-		[ reportParams ]
-	);
-
-	const { primary, hasComparison, isLoading, isFetching, hasData, isError, error, refetch } =
-		useReportOrderAttribution( params );
-
-	const isInitialLoading = isLoading && ! hasData;
-	const isRefetching = isFetching && hasData;
-
-	const chartData = useMemo( () => buildSalesByUtmData( primary.data ), [ primary.data ] );
-	const legendLabels = useMemo( () => formatLegendLabels( reportParams ), [ reportParams ] );
-
-	const hasError = useWidgetError( isError, error, refetch );
-	if ( hasError ) {
-		return null;
-	}
-
-	if ( isInitialLoading ) {
-		return <WidgetLoadingOverlay />;
-	}
-
-	return (
-		<>
-			<LeaderboardChart
-				data={ chartData }
-				withComparison={ hasComparison }
-				legendLabels={ legendLabels }
-				emptyStateIcon={ megaphone }
-				style={
-					{
-						'--a8c--charts--leaderboard--bar--border-radius': '0 1px 1px 0',
-					} as CSSProperties
-				}
-			/>
-			{ isRefetching && <WidgetLoadingOverlay /> }
-		</>
-	);
-}
-
 /**
  * Sales by UTM campaign widget.
  *
  * WidgetRoot provides the query client, chart theme, and resolved report params;
- * this render module fetches the order-attribution report and renders the
- * campaign leaderboard.
+ * the shared SalesByUtmWidget renders the campaign leaderboard.
  */
 export default function SalesByUtmCampaignRender( {
 	attributes = {},
@@ -89,7 +34,7 @@ export default function SalesByUtmCampaignRender( {
 }: SalesByUtmCampaignRenderProps ) {
 	return (
 		<WidgetRoot attributes={ attributes } setError={ setError } options={ { from: '/' } }>
-			<SalesByUtmCampaignWidget />
+			<SalesByUtmWidget view="campaign" />
 		</WidgetRoot>
 	);
 }
