+++ projects/packages/premium-analytics/widgets/sales-by-utm-channel/render.tsx
@@ -1,19 +1,12 @@
 /**
  * External dependencies
  */
-import { useReportOrderAttribution } from '@jetpack-premium-analytics/data';
-import { channel } from '@jetpack-premium-analytics/icons';
 import {
-	LeaderboardChart,
-	WidgetLoadingOverlay,
+	SalesByUtmWidget,
 	WidgetRoot,
-	buildSalesByUtmData,
-	formatLegendLabels,
-	useWidgetError,
-	useWidgetRootContext,
 	type ReportParamsFieldAttributes,
 } from '@jetpack-premium-analytics/widgets-toolkit';
-import { useMemo, type ComponentProps, type CSSProperties } from 'react';
+import { type ComponentProps } from 'react';
 /**
  * Internal dependencies
  */
@@ -29,59 +22,11 @@ type SalesByUtmChannelRenderProps = WidgetRenderProps< SalesByUtmChannelRenderAt
 	setError?: ComponentProps< typeof WidgetRoot >[ 'setError' ];
 };
 
-function SalesByUtmChannelWidget() {
-	const { reportParams } = useWidgetRootContext();
-
-	const params = useMemo(
-		() => ( {
-			...reportParams,
-			view: 'channel' as const,
-		} ),
-		[ reportParams ]
-	);
-
-	const { primary, hasComparison, isLoading, isFetching, hasData, isError, error, refetch } =
-		useReportOrderAttribution( params );
-
-	const isInitialLoading = isLoading && ! hasData;
-	const isRefetching = isFetching && hasData;
-
-	const chartData = useMemo( () => buildSalesByUtmData( primary.data ), [ primary.data ] );
-	const legendLabels = useMemo( () => formatLegendLabels( reportParams ), [ reportParams ] );
-
-	const hasError = useWidgetError( isError, error, refetch );
-	if ( hasError ) {
-		return null;
-	}
-
-	if ( isInitialLoading ) {
-		return <WidgetLoadingOverlay />;
-	}
-
-	return (
-		<>
-			<LeaderboardChart
-				data={ chartData }
-				withComparison={ hasComparison }
-				legendLabels={ legendLabels }
-				emptyStateIcon={ channel }
-				style={
-					{
-						'--a8c--charts--leaderboard--bar--border-radius': '0 1px 1px 0',
-					} as CSSProperties
-				}
-			/>
-			{ isRefetching && <WidgetLoadingOverlay /> }
-		</>
-	);
-}
-
 /**
  * Sales by UTM channel widget.
  *
  * WidgetRoot provides the query client, chart theme, and resolved report params;
- * this render module fetches the order-attribution report and renders the
- * channel leaderboard.
+ * the shared SalesByUtmWidget renders the channel leaderboard.
  */
 export default function SalesByUtmChannelRender( {
 	attributes = {},
@@ -89,7 +34,7 @@ export default function SalesByUtmChannelRender( {
 }: SalesByUtmChannelRenderProps ) {
 	return (
 		<WidgetRoot attributes={ attributes } setError={ setError } options={ { from: '/' } }>
-			<SalesByUtmChannelWidget />
+			<SalesByUtmWidget view="channel" />
 		</WidgetRoot>
 	);
 }
